Rename BBH measurements file
