+++ src/ControlSystem/ControlErrors/Expansion.hpp
@@ -6,7 +6,7 @@
 #include <cstddef>
 #include <pup.h>
 
-#include "ControlSystem/ApparentHorizons/Measurements.hpp"
+#include "ControlSystem/ApparentHorizons/BothHorizons.hpp"
 #include "ControlSystem/Protocols/ControlError.hpp"
 #include "ControlSystem/Tags.hpp"
 #include "DataStructures/DataBox/DataBox.hpp"
