+++ src/ControlSystem/Systems/Expansion.hpp
@@ -9,7 +9,7 @@
 #include <string>
 
 #include "ApparentHorizons/Tags.hpp"
-#include "ControlSystem/ApparentHorizons/Measurements.hpp"
+#include "ControlSystem/ApparentHorizons/BothHorizons.hpp"
 #include "ControlSystem/Component.hpp"
 #include "ControlSystem/ControlErrors/Expansion.hpp"
 #include "ControlSystem/DataVectorHelpers.hpp"
