+++ src/ControlSystem/Systems/Rotation.hpp
@@ -9,7 +9,7 @@
 #include <string>
 
 #include "ApparentHorizons/Tags.hpp"
-#include "ControlSystem/ApparentHorizons/Measurements.hpp"
+#include "ControlSystem/ApparentHorizons/BothHorizons.hpp"
 #include "ControlSystem/Component.hpp"
 #include "ControlSystem/ControlErrors/Rotation.hpp"
 #include "ControlSystem/DataVectorHelpers.hpp"
